+++ react/InputFieldContainer.js
@@ -1,14 +1,16 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
+import reduce from 'lodash/reduce'
+import msk from 'msk'
+
 import AddressShapeWithValidation from './propTypes/AddressShapeWithValidation'
 import {
   getListOfOptions,
   hasOptions,
   getDependentFields,
 } from './selectors/fields'
-import reduce from 'lodash/reduce'
-import msk from 'msk'
 import pureInputField from './pureInputField'
+import { addressContextPropTypes } from './addressContainerContext'
 
 class InputFieldContainer extends Component {
   clearDependentFields(address, dependentFields) {
@@ -18,6 +20,7 @@ class InputFieldContainer extends Component {
       address,
       (cleanAddress, addressField, prop) => {
         const isDependentField = dependentFields.indexOf(prop) !== -1
+
         return isDependentField
           ? {
               ...cleanAddress,
@@ -28,7 +31,7 @@ class InputFieldContainer extends Component {
             }
           : cleanAddress
       },
-      {},
+      {}
     )
   }
 
@@ -36,7 +39,7 @@ class InputFieldContainer extends Component {
     const { field, address, rules, onChangeAddress } = this.props
     const dependentFields = getDependentFields(field.name, rules)
 
-    return value => {
+    return (value) => {
       const clearedFields = this.clearDependentFields(address, dependentFields)
 
       onChangeAddress({
@@ -58,9 +61,9 @@ class InputFieldContainer extends Component {
     return () => {
       onChangeAddress({
         number: {
-          ...address['number'],
-          value: address['number'].disabled ? null : labelNotApplicable,
-          disabled: !address['number'].disabled,
+          ...address.number,
+          value: address.number.disabled ? null : labelNotApplicable,
+          disabled: !address.number.disabled,
         },
       })
     }
@@ -87,7 +90,7 @@ class InputFieldContainer extends Component {
     }
   }
 
-  inputRef = el => {
+  inputRef = (el) => {
     this.el = el
   }
 
@@ -102,6 +105,7 @@ class InputFieldContainer extends Component {
   addFocusIfNeeded() {
     const { address, field, onChangeAddress } = this.props
     const fieldValue = address[field.name]
+
     if (this.el && typeof this.el.focus === 'function' && fieldValue.focus) {
       this.el.focus()
       onChangeAddress({
@@ -125,6 +129,7 @@ class InputFieldContainer extends Component {
       onSubmit,
       submitLabel,
       rules,
+      fieldsStyleRules,
       shouldShowNumberKeyboard,
     } = this.props
 
@@ -148,6 +153,7 @@ class InputFieldContainer extends Component {
       <Input
         Button={Button}
         address={address}
+        fieldsStyleRules={fieldsStyleRules}
         loading={loading}
         field={field}
         autoFocus={autoFocus}
@@ -171,6 +177,7 @@ InputFieldContainer.propTypes = {
 }
 
 InputFieldContainer.propTypes = {
+  ...addressContextPropTypes,
   Button: PropTypes.func,
   Input: PropTypes.func.isRequired,
   loading: PropTypes.bool,