+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- `AddressContainer.fieldsStyleRules` prop should be passed down by every `AddressForm` children.
+
 ## [4.11.0] - 2021-10-26
 
 ### Added